refactor(payment-order): split container.go to stay under 800-line file limit
Extract client creation and helper functions into container_clients.go
to comply with Service Conventions CI check (800-line file limit).
container.go: 604 lines, container_clients.go: 271 lines.
